move compiler code to cli/js/compiler/
