+++ cli/js/deno.ts
@@ -11,7 +11,7 @@ export {
 export { build, OperatingSystem, Arch } from "./build.ts";
 export { chmodSync, chmod } from "./ops/fs/chmod.ts";
 export { chownSync, chown } from "./ops/fs/chown.ts";
-export { transpileOnly, compile, bundle } from "./compiler_api.ts";
+export { transpileOnly, compile, bundle } from "./compiler/api.ts";
 export { inspect } from "./console.ts";
 export { copyFileSync, copyFile } from "./ops/fs/copy_file.ts";
 export {
